+++ src/main/java/Main.java
@@ -4,3 +4,5 @@ public class Main {
         new Runner().run();
     }
 }
+
+
