+++ src/main/java/domain/board/MovePath.java
@@ -1,6 +1,9 @@
-package domain;
+package domain.board;
 
 import java.util.List;
 
 public record MovePath(List<Direction> steps) {}
+
+
+
