+++ src/main/java/domain/board/Route.java
@@ -1,5 +1,8 @@
-package domain;
+package domain.board;
 
 import java.util.List;
 
 public record Route(Position startPos, Position endPos, List<Position> intermeidateNodes) {}
+
+
+
