+++ src/main/java/domain/piece/TeamColor.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.piece;
 
 public enum TeamColor {
     CHO,
@@ -11,3 +11,6 @@ public enum TeamColor {
         return "한";
     }
 }
+
+
+
