+++ src/main/java/strategy/formation/OuterFormationStrategy.java
@@ -1,9 +1,9 @@
 package strategy.formation;
 
-import domain.Piece;
-import domain.PieceType;
-import domain.Position;
-import domain.TeamColor;
+import domain.piece.Piece;
+import domain.piece.PieceType;
+import domain.board.Position;
+import domain.piece.TeamColor;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -25,3 +25,5 @@ public class OuterFormationStrategy extends InitialFormationStrategy {
         return formation;
     }
 }
+
+
