+++ src/main/java/strategy/move/ElephantMoveStrategy.java
@@ -1,8 +1,8 @@
 package strategy.move;
 
-import domain.Direction;
-import domain.MovePath;
-import domain.TeamColor;
+import domain.board.Direction;
+import domain.board.MovePath;
+import domain.piece.TeamColor;
 import java.util.List;
 
 public class ElephantMoveStrategy implements MoveStrategy {
@@ -24,3 +24,5 @@ public class ElephantMoveStrategy implements MoveStrategy {
     }
 
 }
+
+
