+++ src/main/java/strategy/move/PawnMoveStrategy.java
@@ -1,8 +1,8 @@
 package strategy.move;
 
-import domain.Direction;
-import domain.MovePath;
-import domain.TeamColor;
+import domain.board.Direction;
+import domain.board.MovePath;
+import domain.piece.TeamColor;
 import java.util.List;
 
 public class PawnMoveStrategy implements MoveStrategy{
@@ -25,3 +25,5 @@ public class PawnMoveStrategy implements MoveStrategy{
         );
     }
 }
+
+
