+++ src/test/java/domain/piece/PieceTest.java
@@ -1,5 +1,7 @@
-package domain;
+package domain.piece;
 
+import domain.board.Position;
+import domain.board.Route;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.Nested;
@@ -13,7 +15,7 @@ class PieceTest {
     class 전략위임 {
         @Test
         void 기물은_자기_전략으로_후보_경로를_생성한다() {
-            Piece piece = Piece.of(TeamColor.CHO, PieceType.PAWN);
+            final Piece piece = Piece.of(TeamColor.CHO, PieceType.PAWN);
 
             assertThat(piece.makeRoutes(Position.of(2, 3))).containsExactlyInAnyOrder(
                     new Route(Position.of(2, 3), Position.of(1, 3), List.of()),
@@ -24,10 +26,10 @@ class PieceTest {
 
         @Test
         void 기물은_자기_전략으로_이동_가능_여부를_판단한다() {
-            Piece piece = Piece.of(TeamColor.CHO, PieceType.PAWN);
-            Route route = new Route(Position.of(2, 3), Position.of(1, 3), List.of());
+            final Piece piece = Piece.of(TeamColor.CHO, PieceType.PAWN);
+            final Route route = new Route(Position.of(2, 3), Position.of(1, 3), List.of());
 
-            boolean canMove = piece.canMove(
+            final boolean canMove = piece.canMove(
                     route,
                     List.of(),
                     Optional.of(Piece.of(TeamColor.HAN, PieceType.HORSE))
@@ -37,3 +39,6 @@ class PieceTest {
         }
     }
 }
+
+
+
