+++ src/test/java/strategy/move/CannonMoveStrategyTest.java
@@ -1,12 +1,12 @@
 package strategy.move;
 
-import domain.Direction;
-import domain.MovePath;
-import domain.Piece;
-import domain.PieceType;
-import domain.Position;
-import domain.Route;
-import domain.TeamColor;
+import domain.board.Direction;
+import domain.board.MovePath;
+import domain.piece.Piece;
+import domain.piece.PieceType;
+import domain.board.Position;
+import domain.board.Route;
+import domain.piece.TeamColor;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.Nested;
@@ -169,3 +169,5 @@ public class CannonMoveStrategyTest {
         }
     }
 }
+
+
