+++ src/test/java/strategy/move/MoveStategyTest.java
@@ -6,3 +6,5 @@ public class MoveStategyTest {
 
 
 }
+
+
